# perf(loader): eliminate double parsing