## Proposed changes

Previously, each template was parsed twice during
validation:
1. First in `LoadTemplate()` via `Parser.ParseTemplate()`.
2. Then again in `templates.Parse()` for validation checks.

This caused significant overhead when validating
10,000+ templates in nuclei-templates repo.

### Proof

```console
$ hyperfine --warmup=5 --prepare='sleep 5' './bin/nuclei-dev -validate -duc' './bin/nuclei -validate -duc'
Benchmark 1: ./bin/nuclei-dev -validate -duc
  Time (mean ± σ):     93.662 s ± 41.013 s    [User: 24.995 s, System: 2.811 s]
  Range (min … max):   51.213 s … 169.758 s    10 runs
 
Benchmark 2: ./bin/nuclei -validate -duc
  Time (mean ± σ):      2.490 s ±  0.025 s    [User: 3.163 s, System: 0.301 s]
  Range (min … max):    2.456 s …  2.517 s    10 runs
 
Summary
  ./bin/nuclei -validate -duc ran
   37.62 ± 16.48 times faster than ./bin/nuclei-dev -validate -duc
```

> [!NOTE]
> * `./bin/nuclei-dev` built against #6500.
> * `./bin/nuclei` built against this PR branch.

Key Performance Metrics:

- **37.62x faster** validation (3,762% improvement)
- **Reduced from ~93.7s to ~2.5s** (1.5 minutes down to 2.5 seconds)

## Checklist

<!-- Put an "x" in the boxes that apply. You can also fill these out after creating the PR. If you're unsure about any of them, don't hesitate to ask. We're here to help! This is simply a reminder of what we are going to look for before merging your code. -->

- [ ] Pull request is created against the [dev](https://github.com/projectdiscovery/nuclei/tree/dev) branch
- [ ] All checks passed (lint, unit/integration/regression tests etc.) with my changes
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)